yanka/Webapp/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pypdf
python-docx
boto3
